Edited confirmation Dialog
