+++ app/components/alerts/SnackBar/SnackBar.tsx
@@ -1,18 +1,29 @@
-import {CheckCircle} from '@components/assets/Icon/CheckCircle'
+import {FaCheckCircle} from 'react-icons/fa'
+
+import CloseButton from '@components/form/Button/Close'
 
 const SnackBar = ({
   color,
   message,
+  onClose,
 }: {
   color: string
   message: string
+  onClose: () => void
 }): JSX.Element => (
-  <div
-    className={`animate-bounce absolute flex h-14 w-96 min-w-max bottom-4 left-4 p-3 bg-${color}-600 rounded`}
+  <section
+    className={`absolute h-20 w-96 min-w-max bottom-6 left-4 p-2 bg-${color}-600 rounded`}
   >
-    <CheckCircle />
-    <h1 className='text-xl text-white text-center ml-2'>{message}</h1>
-  </div>
+    <div className='flex flex-col'>
+      <article className='h-1/3 w-full flex justify-end right-0'>
+        <CloseButton onClickAction={onClose} />
+      </article>
+      <article className='flex h-2/3'>
+        <FaCheckCircle />
+        <h3 className='text-xl text-white text-center ml-2'>{message}</h3>
+      </article>
+    </div>
+  </section>
 )
 
 export default SnackBar
